Corrected confirmation prompts to delete actions
Refactored confirmation prompts before deleting stacks and flashcards to prevent accidental deletions. Updated the GeneralHelperService to improve user confirmation messaging and ensure uniformity across the application.

+++ Flashcards/View/Commands/FlashcardsMenu/DeleteFlashcard.cs
@@ -12,25 +12,27 @@ namespace Flashcards.View.Commands.FlashcardsMenu;
 internal sealed class DeleteFlashcard : ICommand
 {
     private readonly IFlashcardsRepository _flashcardsRepository;
-    private readonly IMenuCommandFactory<StackMenuEntries> _stackMenuCommandFactory;
     private readonly IMenuCommandFactory<FlashcardEntries> _flashcardMenuCommandFactory;
 
     public DeleteFlashcard(
-        IFlashcardsRepository flashcardsRepository, 
-        IMenuCommandFactory<StackMenuEntries> stackMenuCommandFactory,
+        IFlashcardsRepository flashcardsRepository,
         IMenuCommandFactory<FlashcardEntries> flashcardMenuCommandFactory
             )
     {
         _flashcardsRepository = flashcardsRepository;
-        _stackMenuCommandFactory = stackMenuCommandFactory;
         _flashcardMenuCommandFactory = flashcardMenuCommandFactory;
     }
 
     public void Execute()
     {
-        StackChooserService.GetStacks(_stackMenuCommandFactory);
         FlashcardHelperService.GetFlashcard(_flashcardMenuCommandFactory);
+
+        var confirmation = GeneralHelperService.AskForConfirmation();
         
+        if (!confirmation)
+        {
+            return;
+        }
         
         var result = _flashcardsRepository.Delete();
